IMP laposte_fr: split tests per files
