+++ roulier/carriers/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from . import laposte_fr
-# from . import chronopost_fr
+#from . import chronopost_fr
 
 # from . import geodis
 # from . import dpd
