+++ roulier/carriers/laposte_fr/tests/__init__.py
@@ -1,1 +1,3 @@
-from . import test_laposte
+from . import test_laposte_basic
+from . import test_laposte_europe
+from . import test_laposte_outside_europe
